+++ tests/fixtures/ephemeral/C.py
@@ -1,5 +1,5 @@
 def configure(context):
-    context.stage("tests.fixtures.ephemeral.B")
+    context.stage("tests.fixtures.ephemeral.A")
 
 def execute(context):
     pass