+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.2.0-dev**
 
+- Change behaviour of ephemeral stages, use as much caching as possible
 - Add option to serialize parallel contexts for profiling
 - Increment pickle protocol to version 4
 - Export pipeline flowchart as json